# doc/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NexusHub V1 API",
    "version": "1.0.0-design",
    "description": "开发前契约，尚无运行服务。全部业务POST，只有CSRF引导为GET。跨字段领域规则见模块文档。"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源部署"
    }
  ],
...
  "paths": {
...
    "/api/admin/users/search": {
      "post": {
        "operationId": "searchUsers",
        "tags": [
          "identity"
        ],
        "summary": "账号分页",
        "description": "权限：user:manage。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPageResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSearchRequest"
              }
            }
          }
        }
      }
    },
    "/api/admin/users/create": {
      "post": {
        "operationId": "createUser",
        "tags": [
          "identity"
        ],
        "summary": "创建账号",
        "description": "权限：user:manage。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "同一用户、同一路径、相同业务动作重试复用，建议UUID"
          }
        ]
      }
    },
    "/api/admin/users/status": {
      "post": {
        "operationId": "changeUserStatus",
        "tags": [
          "identity"
        ],
        "summary": "启停账号",
        "description": "权限：user:manage。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStatusRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "同一用户、同一路径、相同业务动作重试复用，建议UUID"
          }
        ]
      }
    },
    "/api/admin/users/reset-password": {
      "post": {
        "operationId": "resetPassword",
        "tags": [
          "identity"
        ],
        "summary": "重置密码",
        "description": "权限：user:manage。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "同一用户、同一路径、相同业务动作重试复用，建议UUID"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "NEXUSHUB_SESSION"
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN"
      }
    },
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string",
            "maxLength": 128
          },
          "reason": {
            "type": "string",
            "maxLength": 300
          }
        },
        "required": [
          "field",
          "reason"
        ]
      },
      "ErrorData": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "minItems": 0,
            "maxItems": 100
          },
          "currentVersion": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 2147483647
              },
              {
                "type": "null"
              }
            ]
          },
          "retryAfterSeconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 2147483647
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": []
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "integer",
            "minimum": 40000,
            "maximum": 59999
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorData"
              },
              {
                "type": "null"
              }
            ]
          },
          "requestId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "requestId"
        ]
      },
...
      "User": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 19,
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "正整数 BIGINT 的十进制字符串，服务端还校验 <= 9223372036854775807"
          },
          "username": {
            "type": "string",
            "maxLength": 64
          },
          "displayName": {
            "type": "string",
            "maxLength": 64
          },
          "role": {
            "type": "string",
            "enum": [
              "GUEST",
              "USER",
              "ADMIN"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ENABLED",
              "DISABLED"
            ]
          },
          "rowVersion": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2147483647
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 64
            },
            "minItems": 0,
            "maxItems": 100
          }
        },
        "required": [
          "id",
          "username",
          "displayName",
          "role",
          "status",
          "rowVersion",
          "permissions"
        ]
      },
      "CreateUserRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "pattern": "^[a-z0-9][a-z0-9_.-]{2,63}$"
          },
          "displayName": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1
          },
          "password": {
            "type": "string",
            "maxLength": 64,
            "minLength": 12,
            "format": "password"
          },
          "role": {
            "type": "string",
            "enum": [
              "GUEST",
              "USER",
              "ADMIN"
            ]
          }
        },
        "required": [
          "username",
          "displayName",
          "password",
          "role"
        ]
      },
      "UserStatusRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 19,
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "正整数 BIGINT 的十进制字符串，服务端还校验 <= 9223372036854775807"
          },
          "expectedRowVersion": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2147483647
          },
          "status": {
            "type": "string",
            "enum": [
              "ENABLED",
              "DISABLED"
            ]
          }
        },
        "required": [
          "id",
          "expectedRowVersion",
          "status"
        ]
      },
      "ResetPasswordRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 19,
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "正整数 BIGINT 的十进制字符串，服务端还校验 <= 9223372036854775807"
          },
          "expectedRowVersion": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2147483647
          },
          "newPassword": {
            "type": "string",
            "maxLength": 64,
            "minLength": 12,
            "format": "password"
          }
        },
        "required": [
          "id",
          "expectedRowVersion",
          "newPassword"
        ]
      },
      "UserSearchRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pageNum": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000,
            "default": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100,
            "default": 20
          },
          "keyword": {
            "type": "string",
            "maxLength": 64
          },
          "status": {
            "type": "string",
            "enum": [
              "ENABLED",
              "DISABLED"
            ]
          }
        },
        "required": []
      },
      "UserPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "minItems": 0,
            "maxItems": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "pageNum": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          }
        },
        "required": [
          "items",
          "total",
          "pageNum",
          "pageSize"
        ]
      },
...
      "UserResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": 200,
            "type": "integer"
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "data": {
            "$ref": "#/components/schemas/User"
          },
          "requestId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "requestId"
        ]
      },
...
      "UserPageResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": 200,
            "type": "integer"
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "data": {
            "$ref": "#/components/schemas/UserPage"
          },
          "requestId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "requestId"
        ]
      },
...
    }
  }
}